Reconstruct streaming_timeline
